requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
requests>=2.31.0
langchain[openai]
langgraph 
langsmith
typing_extensions>=4.9.0 